Problem: HaLink is a misnomer
Solution:
1. Rename HaLink class to Motr
2. Move the class to hax.motr package
3. Correct the name usage in the code e.g. in function or variable names to avoid misleading
